+++ census/static/js/cr-leaflet.js
@@ -359,7 +359,7 @@ CensusReporter.SummaryLevelLayer = CensusReporter.GeoJSONLayer.extend({
     _defaultGeojsonOptions: {
         onEachFeature: function(feature, layer) {
             // you can wire behavior to each "feature", or place outline.
-            var profileURL = 'http://censusreporter.org/profiles/' + feature.properties.geoid;
+            var profileURL = 'http://censusreporter.org/profiles/' + feature.properties.level + '/' + feature.properties.code + '/';
             layer.bindPopup("<a href='" + profileURL + "'>" + feature.properties.name + "</a>");
             if (this.style && this.mouseoverStyle) {
                 layer.on('mouseover', function() {