+++ census/urls.py
@@ -1,7 +1,7 @@
 from django.conf import settings
-from django.conf.urls import url, patterns, include
+from django.conf.urls import url, include
 from django.contrib import admin
-from django.core.urlresolvers import reverse_lazy
+from django.urls import reverse_lazy
 from django.http import HttpResponse
 from django.views.decorators.cache import cache_page
 from django.views.generic.base import TemplateView, RedirectView
@@ -22,7 +22,7 @@ BLOCK_ROBOTS = getattr(settings, 'BLOCK_ROBOTS', False)
 geo_levels = 'ward|municipality|province|country'
 
 
-urlpatterns = patterns('',
+urlpatterns = [
     url(
         regex   = '^$',
         view    = cache_page(STANDARD_CACHE_TIME)(HomepageView.as_view()),
@@ -223,4 +223,4 @@ urlpatterns = patterns('',
         name    = 'elasticsearch',
     ),
     ## END LOCAL DEV VERSION OF API ##
-)
+]