+++ CHANGES.rst
@@ -11,6 +11,16 @@ Be sure to upgrade to 1.x before upgrading to 2.x.
 * BREAKING: Table management is now done through the Django admin interface. You must upgrade your existing tables.
 * BREAKING: ``get_objects_by_geo`` has been replaced by ``FieldTable.get_rows_for_geo``.
 
+1.2.1 (14 September 2018)
+-----------------------
+
+* FIX warning for template parameters
+
+1.2 (13 September 2018)
+-----------------------
+
+* Update to Django version 1.11 which has long-term support
+
 1.1.1 (21 November 2017)
 -----------------------
 
@@ -31,7 +41,7 @@ This release of Wazimap introduces some backwards incompatible changes.
 * BREAKING: The Geography model has changed. You must run ``python manage.py migrate`` when updating.
 * BREAKING: The method signature of the ``profile_builder`` function has changed. It is now called as ``get_profile(geo, profile_name, request)``. You will need to update your profile methods to work with a ``geo`` object rather than a ``level`` and a ``code``. Use ``geo.geo_level`` and ``geo.geo_code`` instead.
 * BREAKING: Some methods on the ``GeoData`` object now take a ``Geography`` object rather than a ``geo_level`` and a ``geo_code``.
-* BREAKING: The ``geography_data`` setting must include a geography version at the top of the dict. Use ``''`` by default. eg: ``{'': {'country': 'geo/country.geojson'}}``
+* BREAKING: The ``geometry_data`` setting must include a geography version at the top of the dict. Use ``''`` by default. eg: ``{'': {'country': 'geo/country.geojson'}}``
 * Some methods on the ``GeoData`` object now take an optional ``version`` parameter.
 * A new method ``GeoData.get_comparative_geos`` makes it easier to build the profile page for comparative geographies.
 * FIX: load streetmap tiles over https if necessary